Simplify and merge landlord and tenant sections, schema and messages.
