+++ libs/application/templates/rental-agreement/project.json
@@ -21,7 +21,7 @@
     "extract-strings": {
       "executor": "nx:run-commands",
       "options": {
-        "command": "yarn ts-node -P libs/localization/tsconfig.lib.json libs/localization/scripts/extract libs/application/templates/rental-agreement/src/lib/messages.ts"
+        "command": "yarn ts-node -P libs/localization/tsconfig.lib.json libs/localization/scripts/extract 'libs/application/templates/rental-agreement/src/lib/messages/*.ts'"
       }
     },
     "codegen/frontend-client": {
