+++ libs/application/templates/rental-agreement/src/lib/dataSchema.ts
@@ -8,8 +8,7 @@ import {
   RentalHousingCategoryTypes,
   RentalHousingConditionInspector,
 } from '../utils/enums'
-import { landlordInfo } from './schemas/landlordSchema'
-import { tenantInfo } from './schemas/tenantSchema'
+import { tenantInfo, landlordInfo } from './schemas/landlordAndTenantSchema'
 import { registerProperty } from './schemas/propertySearchSchema'
 import { otherFees } from './schemas/otherFeesSchema'
 import { securityDeposit } from './schemas/securityDepositSchema'
