+++ libs/application/templates/rental-agreement/src/lib/messages/index.ts
@@ -1,9 +1,8 @@
 export * from './housing/housingCondition'
 export * from './housing/housingFireProtections'
-export * from './housing/landlordDetails'
+export * from './housing/landlordAndTenantDetails'
 export * from './housing/registerProperty'
 export * from './housing/specialProvisions'
-export * from './housing/tenantDetails'
 export * from './housing/userRole'
 export * from './rentalInfo/otherFees'
 export * from './rentalInfo/rentalAmount'
